changed: Introduce abstract base class for Query and ConsequenceParams
In order to avoid `ConsequenceParams` from extending the `Query`
concrete class, this commit introduces a `QueryBase` abstract base class
for both implementations. This way, `Query` and `ConsequenceParams` are
now two distinct concrete classes that do no inherit from one another.
