+++ algoliasearch-tests/src/test/java/com/algolia/search/integration/common/async/AsyncRulesTest.java
@@ -269,7 +269,8 @@ public abstract class AsyncRulesTest extends AsyncAlgoliaIntegrationTest {
             new AutomaticFacetFilter("products.properties.fbrand"));
 
     ConsequenceParams params =
-        new ConsequenceParams<>().setAutomaticFacetFilters(automaticFacetFilters);
+        new ConsequenceParams<AutomaticFacetFilter>()
+            .setAutomaticFacetFilters(automaticFacetFilters);
 
     params.setQuery(consequenceQuery);
     Consequence consequence = new Consequence().setParams(params);
@@ -306,7 +307,8 @@ public abstract class AsyncRulesTest extends AsyncAlgoliaIntegrationTest {
             new AutomaticFacetFilter("products.properties.fbrand"));
 
     ConsequenceParams params =
-        new ConsequenceParams<>().setAutomaticOptionalFacetFilters(optionalAutomaticFacetFilters);
+        new ConsequenceParams<AutomaticFacetFilter>()
+            .setAutomaticOptionalFacetFilters(optionalAutomaticFacetFilters);
 
     params.setQuery(consequenceQuery);
     Consequence consequence = new Consequence().setParams(params);
